fix: remove null fields from generated config file
Previously, when generating the `.licensarc` config file, the JSON
serialized Config struct was saved as-is. This resulted in unnecessary
null fields being present in the config file.

This commit addresses the above behavior by implementing a mechanism
to remove null fields from the final JSON object before saving it to
`.licensarc.`

Closes #40