fix(cli): invalid type error when missing exclude flag
This PR resolves an error caused by parsing workspace configuration
and merging it with user-provided CLI arguments/flags.
Users had to explicitly add the `exclude` flag or provide it's value
in the licensarc.json file, which is ultimately undesirable.

Now, users are not required to explicitly provide this flag nor add it
to the licensarc.json file.

Closes #66