refactor: rename verify command to check
This commit renames the "verify" command to "check" without introducing breaking changes.

This is achieved by making the "add" command the default and aliasing it to "verify". Now users can invoke the same functionality by using:

- licensa check [OPTIONS]
- licensa verify [OPTIONS]

Why this change?

- the scope of the current functionality
- "check" is more frequently used than "verify", making it easier to adopt

"check" is often associated with validation or ensuring correctness in many systems and tools, whereas "verify" can sometimes imply a formal or more extensive process.

For example, users might expect a "check" command to quickly validate the license state of a workspace without the additional connotations that "verify" carries.